# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Dashboard API
  description: |
    REST API for managing a stock/ETF portfolio.
    Live prices are fetched from Yahoo Finance (supports NSE `.NS`, BSE `.BO`, and US symbols).
  version: 1.0.0
  contact:
    email: mailstosuthir@gmail.com
servers:
  - url: http://localhost:8080/api
    description: Local development
...
paths:
...
  /auth/security-questions:
    get:
      tags: [Auth]
      summary: List supported security questions
      operationId: listSecurityQuestions
      responses:
        "200":
          description: Security-question catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityQuestion"
    put:
      tags: [Auth]
      summary: Replace own security questions
      operationId: updateSecurityQuestions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSecurityQuestionsRequest"
      responses:
        "200":
          description: Updated current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/recover:
    post:
      tags: [Auth]
      summary: Recover account access with security questions
      description: |
        If answers and a new password are omitted, returns the user's three
        question prompts. If answers and a new password are present, checks all
        answers, resets the password, and issues a new session.
      operationId: recover
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoverRequest"
      responses:
        "200":
          description: Recovery question prompts or recovered user
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoverResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "423":
          $ref: "#/components/responses/Locked"
  /auth/me:
    get:
      tags: [Auth]
      summary: Return the current user
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/password:
    put:
      tags: [Auth]
      summary: Change own password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Updated current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/profile:
    put:
      tags: [Auth]
      summary: Change own profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /holdings:
    get:
      tags: [Holdings]
      summary: List all holdings
      operationId: listHoldings
      responses:
        "200":
          description: Array of holdings sorted by script name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holding"

    post:
      tags: [Holdings]
      summary: Add a new holding
      operationId: createHolding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldingInput"
      responses:
        "201":
          description: Created holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
        "400":
          $ref: "#/components/responses/BadRequest"
  /holdings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: MongoDB ObjectID

    get:
      tags: [Holdings]
      summary: Get a single holding
      operationId: getHolding
      responses:
        "200":
          description: Holding found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Holdings]
      summary: Update a holding
      operationId: updateHolding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldingInput"
      responses:
        "200":
          description: Updated holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Holdings]
      summary: Delete a holding
      operationId: deleteHolding
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /market/price:
    get:
      tags: [Market]
      summary: Get live price for any symbol
      operationId: getMarketPrice
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
          description: Yahoo Finance ticker (e.g. TCS.NS, AAPL, GOLDBEES.NS)
          example: TCS.NS
      responses:
        "200":
          description: Current price
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbol:
                    type: string
                  price:
                    type: number
                    format: double
                  currency:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /market/forex:
    get:
      tags: [Market]
      summary: Get forex exchange rate
      operationId: getForexRate
      parameters:
        - name: from
          in: query
          schema:
            type: string
            default: INR
        - name: to
          in: query
          schema:
            type: string
            default: EUR
      responses:
        "200":
          description: Exchange rate
          content:
            application/json:
              schema:
                type: object
                properties:
                  from:
                    type: string
                  to:
                    type: string
                  rate:
                    type: number
                    format: double
                    description: 1 unit of `from` in `to`
  /admin/users:
    get:
      tags: [Admin]
      summary: List manageable users
      operationId: listAdminUsers
      parameters:
        - name: include_hidden
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Users visible to the caller
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/admins:
    get:
      tags: [Admin]
      summary: List admin and super-admin accounts
      operationId: listAdmins
      responses:
        "200":
          description: Admin accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: User ObjectID
    get:
      tags: [Admin]
      summary: Get a manageable user
      operationId: getAdminUser
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin]
      summary: Permanently delete a user and their holdings
      operationId: deleteAdminUser
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/reset-lockout:
    post:
      tags: [Admin]
      summary: Clear recovery lockout for a user
      operationId: resetAdminUserLockout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/hide:
    post:
      tags: [Admin]
      summary: Hide a user
      operationId: hideAdminUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/reactivate:
    post:
      tags: [Admin]
      summary: Reactivate a hidden user
      operationId: reactivateAdminUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/promote:
    post:
      tags: [Admin]
      summary: Promote a user to admin
      operationId: promoteAdminUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/demote:
    post:
      tags: [Admin]
      summary: Demote an admin to user
      operationId: demoteAdminUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/region:
    put:
      tags: [Admin]
      summary: Move a user or admin to another region
      operationId: updateAdminUserRegion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRegionRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/holdings:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Admin]
      summary: List a target user's holdings
      operationId: listAdminUserHoldings
      responses:
        "200":
          description: Target user's holdings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holding"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Admin]
      summary: Create a holding for a target user
      operationId: createAdminUserHolding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldingInput"
      responses:
        "201":
          description: Created holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/holdings/{holding_id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: holding_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Admin]
      summary: Get a target user's holding
      operationId: getAdminUserHolding
      responses:
        "200":
          description: Holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Admin]
      summary: Update a target user's holding
      operationId: updateAdminUserHolding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldingInput"
      responses:
        "200":
          description: Updated holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin]
      summary: Delete a target user's holding
      operationId: deleteAdminUserHolding
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/prices:
    get:
      tags: [Admin]
      summary: Get target user's holdings with prices
      operationId: getAdminUserPrices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Target user's priced holdings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/summary:
    get:
      tags: [Admin]
      summary: Get target user's portfolio summary
      operationId: getAdminUserSummary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Target user's summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  schemas:
...
    SecurityQuestion:
      type: object
      required: [id, prompt]
      properties:
        id:
          type: string
        prompt:
          type: string
    SecurityAnswerInput:
      type: object
      required: [question_id, answer]
      properties:
        question_id:
          type: string
        answer:
          type: string
    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
          description: Display username as typed by the user
        name:
          type: string
        role:
          type: string
          enum: [user, admin, superadmin]
        region:
          type: string
          enum: [india, europe, us]
        disabled:
          type: boolean
        locked:
          type: boolean
        login_failures:
          type: integer
        security_question_failures:
          type: integer
        must_change_password:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
...
    RecoverRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
        answers:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/SecurityAnswerInput"
        new_password:
          type: string
          format: password
    RecoverResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/SecurityQuestion"
        user:
          $ref: "#/components/schemas/AuthUser"
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
    UpdateProfileRequest:
      type: object
      required: [current_password, username, name]
      properties:
        current_password:
          type: string
          format: password
        username:
          type: string
        name:
          type: string
    UpdateSecurityQuestionsRequest:
      type: object
      required: [current_password, security_questions]
      properties:
        current_password:
          type: string
          format: password
        security_questions:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/SecurityAnswerInput"
    UpdateUserRegionRequest:
      type: object
      required: [region]
      properties:
        region:
          type: string
          enum: [india, europe, us]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    Holding:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectID
        script:
          type: string
          description: Display name (e.g. "TCS", "GOLD BEES")
        symbol:
          type: string
          description: Yahoo Finance ticker (e.g. "TCS.NS", "GOLDBEES.NS", "AAPL")
        exchange:
          type: string
          enum: [NSE, BSE, NYSE, NASDAQ, OTHER]
        type:
          type: string
          enum: [stock, etf]
        stocks_owned:
          type: number
          format: double
          description: Current quantity held
        avg_cost_price:
          type: number
          format: double
          description: Average purchase price per share in the holding's currency
        realized_pnl:
          type: number
          format: double
          description: Profit/loss from shares already sold, in the holding's currency
        currency:
          type: string
          enum: [INR, EUR]
          default: INR
          description: Currency in which avg_cost_price and realized_pnl are denominated
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    HoldingInput:
      type: object
      required: [script, exchange, type]
      properties:
        script:
          type: string
        symbol:
          type: string
        exchange:
          type: string
          enum: [NSE, BSE, NYSE, NASDAQ, OTHER]
        type:
          type: string
          enum: [stock, etf]
        stocks_owned:
          type: number
          format: double
          default: 0
        avg_cost_price:
          type: number
          format: double
          default: 0
        realized_pnl:
          type: number
          format: double
          default: 0
        currency:
          type: string
          enum: [INR, EUR]
          default: INR
          description: Currency in which avg_cost_price and realized_pnl are denominated
        notes:
          type: string
    HoldingWithPrice:
      allOf:
        - $ref: "#/components/schemas/Holding"
        - type: object
          properties:
            current_price:
              type: number
              format: double
            cost_price:
              type: number
              format: double
              description: stocks_owned × avg_cost_price
            current_value:
              type: number
              format: double
              description: stocks_owned × current_price
            unrealized_pnl:
              type: number
              format: double
              description: current_value − cost_price
            cost_price_eur:
              type: number
              format: double
            current_value_eur:
              type: number
              format: double
            unrealized_pnl_eur:
              type: number
              format: double
            realized_pnl_eur:
              type: number
              format: double
            price_error:
              type: string
    PricesResponse:
      type: object
      properties:
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/HoldingWithPrice"
        eur_rate:
          type: number
          format: double
          description: Live INR→EUR rate used for conversions
    Summary:
      type: object
      properties:
        total_cost:
          type: number
          format: double
        total_current_value:
          type: number
          format: double
        total_unrealized:
          type: number
          format: double
        total_realized:
          type: number
          format: double
        total_cost_eur:
          type: number
          format: double
        total_current_value_eur:
          type: number
          format: double
        total_unrealized_eur:
          type: number
          format: double
        total_realized_eur:
          type: number
          format: double
        eur_rate:
          type: number
          format: double
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Locked:
      description: Account is locked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: Upstream (Yahoo Finance) error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"